requirements.txt
# VRSEN PubScrape - Advanced Multi-Agent Web Scraping Platform

# Core scraping and browser automation
botasaurus>=4.0.0
selenium>=4.15.0

# Agency Swarm for multi-agent orchestration
agency-swarm>=0.4.0

# Web scraping and parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16

# Data validation and processing
pydantic>=2.5.0
email-validator>=2.1.0

# Rate limiting and delays
ratelimit>=2.2.1
tenacity>=8.2.3

# Configuration and YAML parsing
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Security libraries
cryptography>=41.0.0
bcrypt>=4.1.0

# CSV and data export
pandas>=2.1.0
openpyxl>=3.1.0

# Logging and monitoring
structlog>=23.2.0
coloredlogs>=15.0.1

# Performance monitoring
psutil>=5.9.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Proxy support
requests-toolbelt>=1.0.0
urllib3>=2.1.0

# DNS and network utilities
dnspython>=2.4.0

# OpenAI for AI agents
openai>=1.12.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Additional dependencies for main system
asyncio-throttle>=1.0.2
aiofiles>=23.2.0
aiohttp>=3.9.0
click>=8.1.0
rich>=13.7.0
tabulate>=0.9.0

# Database support (optional)
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0

# Caching
redis>=5.0.0
diskcache>=5.6.0

# Compression
zstandard>=0.22.0